runtime: proposals: Move tests.
Move proposals tests to the ‘tests’ folder.
