+++ runtime-modules/proposals/engine/src/types/mod.rs
@@ -249,13 +249,19 @@ pub trait VotersParameters {
 }
 
 // Calculates quorum, votes threshold, expiration status
-struct ProposalStatusResolution<'a, BlockNumber, ProposerId, Balance, AccountId> {
-    proposal: &'a Proposal<BlockNumber, ProposerId, Balance, AccountId>,
-    now: BlockNumber,
-    votes_count: u32,
-    total_voters_count: u32,
-    approvals: u32,
-    slashes: u32,
+pub(crate) struct ProposalStatusResolution<'a, BlockNumber, ProposerId, Balance, AccountId> {
+    /// Proposal data
+    pub proposal: &'a Proposal<BlockNumber, ProposerId, Balance, AccountId>,
+    /// Current block
+    pub now: BlockNumber,
+    /// Total votes number so far
+    pub votes_count: u32,
+    /// Council size
+    pub total_voters_count: u32,
+    /// Approval votes number
+    pub approvals: u32,
+    /// Slash votes number
+    pub slashes: u32,
 }
 
 impl<'a, BlockNumber, ProposerId, Balance, AccountId>
@@ -271,7 +277,7 @@ where
 
     // Approval quorum reached for the proposal. Compares predefined parameter with actual
     // votes sum divided by total possible votes number.
-    fn is_approval_quorum_reached(&self) -> bool {
+    pub(crate) fn is_approval_quorum_reached(&self) -> bool {
         let actual_votes_fraction =
             Perbill::from_rational_approximation(self.votes_count, self.total_voters_count);
         let approval_quorum_fraction =
@@ -282,7 +288,7 @@ where
 
     // Verifies that approval threshold is still achievable for the proposal.
     // Compares actual approval votes sum with remaining possible votes number.
-    fn is_approval_threshold_achievable(&self) -> bool {
+    pub(crate) fn is_approval_threshold_achievable(&self) -> bool {
         let remaining_votes_count = self.total_voters_count - self.votes_count;
         let possible_approval_votes_fraction = Perbill::from_rational_approximation(
             self.approvals + remaining_votes_count,
@@ -302,7 +308,7 @@ where
 
     // Verifies that slashing threshold is still achievable for the proposal.
     // Compares actual slashing votes sum with remaining possible votes number.
-    fn is_slashing_threshold_achievable(&self) -> bool {
+    pub(crate) fn is_slashing_threshold_achievable(&self) -> bool {
         let remaining_votes_count = self.total_voters_count - self.votes_count;
         let possible_slashing_votes_fraction = Perbill::from_rational_approximation(
             self.slashes + remaining_votes_count,
@@ -468,509 +474,4 @@ pub trait StakingHandler<T: system::Trait + membership::Trait + balances::Trait>
 
     /// Returns the current stake on the account.
     fn current_stake(account_id: &T::AccountId) -> BalanceOf<T>;
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::types::ProposalStatusResolution;
-    use crate::*;
-
-    // Alias introduced for simplicity of changing Proposal exact types.
-    type ProposalObject = Proposal<u64, u64, u64, u64>;
-
-    #[test]
-    fn proposal_voting_period_expired() {
-        let mut proposal = ProposalObject::default();
-
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-
-        assert!(proposal.is_voting_period_expired(4));
-    }
-
-    #[test]
-    fn proposal_voting_period_not_expired() {
-        let mut proposal = ProposalObject::default();
-
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-
-        assert!(!proposal.is_voting_period_expired(3));
-    }
-
-    #[test]
-    fn proposal_grace_period_expired() {
-        let mut proposal = ProposalObject::default();
-
-        proposal.parameters.grace_period = 3;
-        proposal.status = ProposalStatus::finalized(
-            ProposalDecisionStatus::Approved(ApprovedProposalStatus::PendingExecution),
-            0,
-        );
-
-        assert!(proposal.is_grace_period_expired(4));
-    }
-
-    #[test]
-    fn proposal_grace_period_auto_expired() {
-        let mut proposal = ProposalObject::default();
-
-        proposal.parameters.grace_period = 0;
-        proposal.status = ProposalStatus::finalized(
-            ProposalDecisionStatus::Approved(ApprovedProposalStatus::PendingExecution),
-            0,
-        );
-
-        assert!(proposal.is_grace_period_expired(1));
-    }
-
-    #[test]
-    fn proposal_grace_period_not_expired() {
-        let mut proposal = ProposalObject::default();
-
-        proposal.parameters.grace_period = 3;
-
-        assert!(!proposal.is_grace_period_expired(3));
-    }
-
-    #[test]
-    fn proposal_grace_period_not_expired_because_of_not_approved_proposal() {
-        let mut proposal = ProposalObject::default();
-
-        proposal.parameters.grace_period = 3;
-
-        assert!(!proposal.is_grace_period_expired(3));
-    }
-
-    #[test]
-    fn define_proposal_decision_status_returns_expired() {
-        let mut proposal = ProposalObject::default();
-        let now = 5;
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-        proposal.parameters.approval_quorum_percentage = 80;
-        proposal.parameters.approval_threshold_percentage = 40;
-        proposal.parameters.slashing_quorum_percentage = 50;
-        proposal.parameters.slashing_threshold_percentage = 50;
-
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Approve);
-        proposal.voting_results.add_vote(VoteKind::Approve);
-
-        assert_eq!(
-            proposal.voting_results,
-            VotingResults {
-                abstentions: 0,
-                approvals: 2,
-                rejections: 1,
-                slashes: 0,
-            }
-        );
-
-        let expected_proposal_status = proposal.define_proposal_decision_status(5, now);
-        assert_eq!(
-            expected_proposal_status,
-            Some(ProposalDecisionStatus::Expired)
-        );
-    }
-
-    #[test]
-    fn define_proposal_decision_status_returns_approved() {
-        let now = 2;
-        let mut proposal = ProposalObject::default();
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-        proposal.parameters.approval_quorum_percentage = 60;
-        proposal.parameters.slashing_quorum_percentage = 50;
-        proposal.parameters.slashing_threshold_percentage = 50;
-
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Approve);
-        proposal.voting_results.add_vote(VoteKind::Approve);
-        proposal.voting_results.add_vote(VoteKind::Approve);
-
-        assert_eq!(
-            proposal.voting_results,
-            VotingResults {
-                abstentions: 0,
-                approvals: 3,
-                rejections: 1,
-                slashes: 0,
-            }
-        );
-
-        let expected_proposal_status = proposal.define_proposal_decision_status(5, now);
-        assert_eq!(
-            expected_proposal_status,
-            Some(ProposalDecisionStatus::Approved(
-                ApprovedProposalStatus::PendingExecution
-            ))
-        );
-    }
-
-    #[test]
-    fn define_proposal_decision_status_returns_rejected() {
-        let mut proposal = ProposalObject::default();
-        let now = 2;
-
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-        proposal.parameters.approval_quorum_percentage = 50;
-        proposal.parameters.approval_threshold_percentage = 51;
-        proposal.parameters.slashing_quorum_percentage = 50;
-        proposal.parameters.slashing_threshold_percentage = 50;
-
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Abstain);
-        proposal.voting_results.add_vote(VoteKind::Approve);
-
-        assert_eq!(
-            proposal.voting_results,
-            VotingResults {
-                abstentions: 1,
-                approvals: 1,
-                rejections: 2,
-                slashes: 0,
-            }
-        );
-
-        let expected_proposal_status = proposal.define_proposal_decision_status(4, now);
-        assert_eq!(
-            expected_proposal_status,
-            Some(ProposalDecisionStatus::Rejected)
-        );
-    }
-
-    #[test]
-    fn define_proposal_decision_status_returns_slashed() {
-        let mut proposal = ProposalObject::default();
-        let now = 2;
-
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-        proposal.parameters.approval_quorum_percentage = 50;
-        proposal.parameters.approval_threshold_percentage = 50;
-        proposal.parameters.slashing_quorum_percentage = 50;
-        proposal.parameters.slashing_threshold_percentage = 50;
-
-        proposal.voting_results.add_vote(VoteKind::Slash);
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Abstain);
-        proposal.voting_results.add_vote(VoteKind::Slash);
-
-        assert_eq!(
-            proposal.voting_results,
-            VotingResults {
-                abstentions: 1,
-                approvals: 0,
-                rejections: 1,
-                slashes: 2,
-            }
-        );
-
-        let expected_proposal_status = proposal.define_proposal_decision_status(4, now);
-        assert_eq!(
-            expected_proposal_status,
-            Some(ProposalDecisionStatus::Slashed)
-        );
-    }
-
-    #[test]
-    fn define_proposal_decision_status_returns_none() {
-        let mut proposal = ProposalObject::default();
-        let now = 2;
-
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-        proposal.parameters.approval_quorum_percentage = 60;
-        proposal.parameters.slashing_quorum_percentage = 50;
-
-        proposal.voting_results.add_vote(VoteKind::Abstain);
-        assert_eq!(
-            proposal.voting_results,
-            VotingResults {
-                abstentions: 1,
-                approvals: 0,
-                rejections: 0,
-                slashes: 0,
-            }
-        );
-
-        let expected_proposal_status = proposal.define_proposal_decision_status(5, now);
-        assert_eq!(expected_proposal_status, None);
-    }
-
-    #[test]
-    fn define_proposal_decision_status_returns_approved_before_slashing_before_rejection() {
-        let mut proposal = ProposalObject::default();
-        let now = 2;
-
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-        proposal.parameters.approval_quorum_percentage = 50;
-        proposal.parameters.approval_threshold_percentage = 30;
-        proposal.parameters.slashing_quorum_percentage = 50;
-        proposal.parameters.slashing_threshold_percentage = 30;
-
-        proposal.voting_results.add_vote(VoteKind::Approve);
-        proposal.voting_results.add_vote(VoteKind::Approve);
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Slash);
-        proposal.voting_results.add_vote(VoteKind::Slash);
-
-        assert_eq!(
-            proposal.voting_results,
-            VotingResults {
-                abstentions: 0,
-                approvals: 2,
-                rejections: 2,
-                slashes: 2,
-            }
-        );
-
-        let expected_proposal_status = proposal.define_proposal_decision_status(6, now);
-
-        assert_eq!(
-            expected_proposal_status,
-            Some(ProposalDecisionStatus::Approved(
-                ApprovedProposalStatus::PendingExecution
-            ))
-        );
-    }
-
-    #[test]
-    fn define_proposal_decision_status_returns_slashed_before_rejection() {
-        let mut proposal = ProposalObject::default();
-        let now = 2;
-
-        proposal.activated_at = 1;
-        proposal.parameters.voting_period = 3;
-        proposal.parameters.approval_quorum_percentage = 50;
-        proposal.parameters.approval_threshold_percentage = 30;
-        proposal.parameters.slashing_quorum_percentage = 50;
-        proposal.parameters.slashing_threshold_percentage = 30;
-
-        proposal.voting_results.add_vote(VoteKind::Abstain);
-        proposal.voting_results.add_vote(VoteKind::Approve);
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Reject);
-        proposal.voting_results.add_vote(VoteKind::Slash);
-        proposal.voting_results.add_vote(VoteKind::Slash);
-
-        assert_eq!(
-            proposal.voting_results,
-            VotingResults {
-                abstentions: 1,
-                approvals: 1,
-                rejections: 2,
-                slashes: 2,
-            }
-        );
-
-        let expected_proposal_status = proposal.define_proposal_decision_status(6, now);
-
-        assert_eq!(
-            expected_proposal_status,
-            Some(ProposalDecisionStatus::Slashed)
-        );
-    }
-
-    #[test]
-    fn proposal_status_resolution_approval_quorum_works_correctly() {
-        let no_approval_quorum_proposal: Proposal<u64, u64, u64, u64> = Proposal {
-            parameters: ProposalParameters {
-                approval_quorum_percentage: 63,
-                slashing_threshold_percentage: 63,
-                ..ProposalParameters::default()
-            },
-            ..Proposal::default()
-        };
-        let no_approval_proposal_status_resolution = ProposalStatusResolution {
-            proposal: &no_approval_quorum_proposal,
-            now: 20,
-            votes_count: 314,
-            total_voters_count: 500,
-            approvals: 3,
-            slashes: 3,
-        };
-
-        assert!(!no_approval_proposal_status_resolution.is_approval_quorum_reached());
-
-        let approval_quorum_proposal_status_resolution = ProposalStatusResolution {
-            votes_count: 315,
-            ..no_approval_proposal_status_resolution
-        };
-
-        assert!(approval_quorum_proposal_status_resolution.is_approval_quorum_reached());
-    }
-
-    #[test]
-    fn proposal_status_resolution_slashing_quorum_works_correctly() {
-        let no_slashing_quorum_proposal: Proposal<u64, u64, u64, u64> = Proposal {
-            parameters: ProposalParameters {
-                approval_quorum_percentage: 63,
-                slashing_quorum_percentage: 63,
-                ..ProposalParameters::default()
-            },
-            ..Proposal::default()
-        };
-        let no_slashing_proposal_status_resolution = ProposalStatusResolution {
-            proposal: &no_slashing_quorum_proposal,
-            now: 20,
-            votes_count: 314,
-            total_voters_count: 500,
-            approvals: 3,
-            slashes: 3,
-        };
-
-        assert!(!no_slashing_proposal_status_resolution.is_slashing_quorum_reached());
-
-        let slashing_quorum_proposal_status_resolution = ProposalStatusResolution {
-            votes_count: 315,
-            ..no_slashing_proposal_status_resolution
-        };
-
-        assert!(slashing_quorum_proposal_status_resolution.is_slashing_quorum_reached());
-    }
-
-    #[test]
-    fn proposal_status_resolution_approval_threshold_works_correctly() {
-        let no_approval_threshold_proposal: Proposal<u64, u64, u64, u64> = Proposal {
-            parameters: ProposalParameters {
-                slashing_threshold_percentage: 63,
-                approval_threshold_percentage: 63,
-                ..ProposalParameters::default()
-            },
-            ..Proposal::default()
-        };
-        let no_approval_proposal_status_resolution = ProposalStatusResolution {
-            proposal: &no_approval_threshold_proposal,
-            now: 20,
-            votes_count: 500,
-            total_voters_count: 600,
-            approvals: 314,
-            slashes: 3,
-        };
-
-        assert!(!no_approval_proposal_status_resolution.is_approval_threshold_reached());
-
-        let approval_threshold_proposal_status_resolution = ProposalStatusResolution {
-            approvals: 315,
-            ..no_approval_proposal_status_resolution
-        };
-
-        assert!(approval_threshold_proposal_status_resolution.is_approval_threshold_reached());
-    }
-
-    #[test]
-    fn proposal_status_resolution_slashing_threshold_works_correctly() {
-        let no_slashing_threshold_proposal: Proposal<u64, u64, u64, u64> = Proposal {
-            parameters: ProposalParameters {
-                slashing_threshold_percentage: 63,
-                approval_threshold_percentage: 63,
-                ..ProposalParameters::default()
-            },
-            ..Proposal::default()
-        };
-        let no_slashing_proposal_status_resolution = ProposalStatusResolution {
-            proposal: &no_slashing_threshold_proposal,
-            now: 20,
-            votes_count: 500,
-            total_voters_count: 600,
-            approvals: 3,
-            slashes: 314,
-        };
-
-        assert!(!no_slashing_proposal_status_resolution.is_slashing_threshold_reached());
-
-        let slashing_threshold_proposal_status_resolution = ProposalStatusResolution {
-            slashes: 315,
-            ..no_slashing_proposal_status_resolution
-        };
-
-        assert!(slashing_threshold_proposal_status_resolution.is_slashing_threshold_reached());
-    }
-
-    #[test]
-    fn proposal_exact_execution_block_reached() {
-        let mut proposal = ProposalObject::default();
-
-        proposal.exact_execution_block = None;
-        assert!(proposal.is_execution_block_reached_or_not_set(3));
-
-        proposal.exact_execution_block = Some(3);
-        assert!(proposal.is_execution_block_reached_or_not_set(3));
-    }
-
-    #[test]
-    fn proposal_exact_execution_block_not_reached() {
-        let mut proposal = ProposalObject::default();
-
-        proposal.exact_execution_block = Some(3);
-        assert!(!proposal.is_execution_block_reached_or_not_set(2));
-    }
-
-    #[test]
-    fn proposal_status_resolution_is_approval_achievable_works_correctly() {
-        let approval_threshold_proposal: Proposal<u64, u64, u64, u64> = Proposal {
-            parameters: ProposalParameters {
-                slashing_threshold_percentage: 50,
-                approval_threshold_percentage: 50,
-                ..ProposalParameters::default()
-            },
-            ..Proposal::default()
-        };
-        let not_achievable_proposal_status_resolution = ProposalStatusResolution {
-            proposal: &approval_threshold_proposal,
-            now: 20,
-            votes_count: 302,
-            total_voters_count: 600,
-            approvals: 1,
-            slashes: 0,
-        };
-
-        assert!(!not_achievable_proposal_status_resolution.is_approval_threshold_achievable());
-        assert!(not_achievable_proposal_status_resolution.is_rejection_imminent());
-
-        let approval_threshold_achievable_resolution = ProposalStatusResolution {
-            approvals: 2,
-            ..not_achievable_proposal_status_resolution
-        };
-
-        assert!(approval_threshold_achievable_resolution.is_approval_threshold_achievable());
-        assert!(!approval_threshold_achievable_resolution.is_rejection_imminent());
-    }
-
-    #[test]
-    fn proposal_status_resolution_is_slashing_achievable_works_correctly() {
-        let slashing_threshold_proposal: Proposal<u64, u64, u64, u64> = Proposal {
-            parameters: ProposalParameters {
-                slashing_threshold_percentage: 50,
-                approval_threshold_percentage: 50,
-                ..ProposalParameters::default()
-            },
-            ..Proposal::default()
-        };
-        let not_achievable_proposal_status_resolution = ProposalStatusResolution {
-            proposal: &slashing_threshold_proposal,
-            now: 20,
-            votes_count: 302,
-            total_voters_count: 600,
-            approvals: 0,
-            slashes: 1,
-        };
-
-        assert!(!not_achievable_proposal_status_resolution.is_approval_threshold_achievable());
-        assert!(not_achievable_proposal_status_resolution.is_rejection_imminent());
-
-        let slashing_threshold_achievable_resolution = ProposalStatusResolution {
-            slashes: 2,
-            ..not_achievable_proposal_status_resolution
-        };
-
-        assert!(slashing_threshold_achievable_resolution.is_slashing_threshold_achievable());
-        assert!(!slashing_threshold_achievable_resolution.is_rejection_imminent());
-    }
 }
